% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dq.R
\docType{package}
\name{dq}
\alias{dq}
\alias{dq-package}
\title{The dq Package}
\description{
Tools for Data Quality.
}
\section{Functions}{


\code{\link{dq_univariate}}

\code{\link{dq_pairwise}}

\code{\link{dq_pca}}

\code{\link{dq_multivariate}}

The various methods listed on the \code{\link{plot_dq}} help page.
}

