% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{dq_pca}
\alias{dq_pca}
\title{Get the eigenvalues of PCA}
\usage{
dq_pca(dat)
}
\arguments{
\item{dat}{The input data set.}
}
\value{
An object of class "dq_pca": a numeric vector containing eigenvalues.
}
\description{
Get the eigenvalues of PCA
}
