% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dq}
\alias{plot_dq}
\alias{plot.dq_univariate}
\alias{plot.dq_multivariate}
\alias{plot.dq_pca}
\title{Plot \code{dq} objects}
\usage{
\method{plot}{dq_univariate}(x, variable, data, ...)

\method{plot}{dq_multivariate}(x, cutoff = 0.05, ...)

\method{plot}{dq_pca}(x, ...)
}
\arguments{
\item{x}{An object}

\item{variable}{Character string: which variable to plot?}

\item{data}{The original data.frame used in \code{\link{dq_univariate}}.}

\item{...}{Other arguments.}

\item{cutoff}{The cutoff for outliers.}
}
\description{
Plot \code{dq} objects
}
