% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{dq_multivariate}
\alias{dq_multivariate}
\alias{format.dq_multivariate}
\title{Compute multivariate quality metrics}
\usage{
dq_multivariate(dat, mc.cores = getOption("mc.cores", 8L))

\method{format}{dq_multivariate}(x, digits.pval = 4, ...)
}
\arguments{
\item{dat}{The input data set}

\item{mc.cores}{See \code{\link[parallel]{mclapply}}}

\item{x}{An R object}

\item{...}{Other arguments.}

\item{digits, digits.pval}{How many digits to print}
}
\description{
Compute multivariate quality metrics
}
