% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{dq_pairwise}
\alias{dq_pairwise}
\alias{format.dq_pairwise}
\title{Compute pairwise quality metrics}
\usage{
dq_pairwise(dat)

\method{format}{dq_pairwise}(x, digits = 3, ...)
}
\arguments{
\item{dat}{The input data set}

\item{x}{An R object}

\item{digits}{How many digits to print}

\item{...}{Other arguments.}
}
\value{
An object of class "dq_pairwise".
}
\description{
Compute pairwise quality metrics
}
