% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{dq_univariate}
\alias{dq_univariate}
\alias{format.dq_univariate}
\title{Compute univariate quality metrics}
\usage{
dq_univariate(dat, cutoff = 0.05)

\method{format}{dq_univariate}(x, digits = 3, digits.pct = 1, ...)
}
\arguments{
\item{dat}{The input data set}

\item{cutoff}{The cutoff to determine outliers.}

\item{x}{An R object}

\item{digits, digits.pct}{How many digits to print}

\item{...}{Other arguments.}
}
\description{
Compute univariate quality metrics
}
\details{
Make sure you set.seed before you run this function to get consistent results.
}
